configs refactor part 3/3: registry-based vision extractor hooks
Mirrors part 2/3 for the vision side: replaces the 185-line
_extract_vision_config switch with a per-model hook registry. The
vision dispatcher carries a bit more logic than audio because vision
also lifts a fixed set of shared fields (image_token_id, mrope_section,
spatial_merge_size, ...) up to the top-level of the returned dict.

New per_model files:
  _vision_default.py        # canonical HF vision_config + shared bits
  _phi4mm_vision.py         # hard-coded SigLIP encoder dims
  _hunyuan_vl_mot_vision.py # InternViT-style ViT, flat config.json
  _internvl_vision.py       # default image_token_id for InternVL chain

After this PR, every model-type switch in the original
_extract_*_config functions has been moved into discoverable per-model
files. New audio- or vision-capable models add a single file under
src/mobius/_configs/per_model/ instead of editing _base.py.

Tests:
  pytest src/ tests/build_graph_test.py tests/cli_test.py -n auto
    2769 passed, 41 skipped
  pytest tests/arch_validation_test.py -k 'phi4mm or hunyuan_vl_mot
                                            or qwen2_vl or qwen3_vl
                                            or gemma4'
    30 passed
Ruff: clean.
