+++ src/mobius/_configs/_base.py
@@ -264,189 +264,14 @@ def _extract_mrope_fields(config) -> dict:
 def _extract_vision_config(config, parent_config, model_type: str) -> dict:
     """Extract vision sub-config from a HuggingFace config.
 
-    Builds a :class:`VisionConfig` (if vision fields are found) and
-    returns a dict of options to merge into :class:`ArchitectureConfig`
-    kwargs.
+    Thin shim that delegates to the per-model registry. The actual
+    hooks live under :mod:`mobius._configs.per_model` and are
+    registered with :mod:`mobius._configs._extractors` at import time.
     """
-    # Use parent_config (composite config) when available
-    vision_source = parent_config or config
-    hf_vision_config = getattr(vision_source, "vision_config", None)
-    if hf_vision_config is None:
-        hf_vision_config = getattr(config, "vision_config", None)
+    from mobius._configs import per_model  # noqa: F401  - side-effect import
+    from mobius._configs._extractors import extract_vision_config as _dispatch
 
-    vision_fields: dict = {}
-    if hf_vision_config is not None:
-        vc = (
-            hf_vision_config
-            if not isinstance(hf_vision_config, dict)
-            else type("VC", (), hf_vision_config)()
-        )
-        # Qwen2-VL uses ``embed_dim`` for the actual block hidden size and
-        # ``hidden_size`` for the projection output.  When ``embed_dim``
-        # exists and ``out_hidden_size`` does not, remap the fields so the
-        # block dimension is used as ``hidden_size``.
-        _embed_dim = getattr(vc, "embed_dim", None)
-        _hf_hidden = getattr(vc, "hidden_size", None)
-        _out_hidden = getattr(vc, "out_hidden_size", None)
-        if _embed_dim is not None and _out_hidden is None and _embed_dim != _hf_hidden:
-            _vis_hidden = _embed_dim
-            _vis_out_hidden = _hf_hidden
-        else:
-            _vis_hidden = _hf_hidden
-            _vis_out_hidden = _out_hidden
-
-        _intermediate = getattr(vc, "intermediate_size", None)
-        if _intermediate is None:
-            _mlp_ratio = getattr(vc, "mlp_ratio", None)
-            # _vis_hidden may be a list for multi-stage models (Florence2)
-            _scalar_hidden = _first(_vis_hidden) if _vis_hidden is not None else None
-            if _mlp_ratio is not None and _scalar_hidden is not None:
-                _intermediate = int(_scalar_hidden * _mlp_ratio)
-
-        vision_fields.update(
-            hidden_size=_vis_hidden,
-            intermediate_size=_intermediate,
-            num_hidden_layers=(
-                getattr(vc, "num_hidden_layers", None) or getattr(vc, "depth", None)
-            ),
-            num_attention_heads=(
-                getattr(vc, "num_attention_heads", None)
-                or getattr(vc, "num_heads", None)
-                or getattr(vc, "attention_heads", None)
-            ),
-            image_size=getattr(vc, "image_size", None),
-            patch_size=getattr(vc, "patch_size", None),
-            norm_eps=_first_not_none(
-                getattr(vc, "layer_norm_eps", None),
-                getattr(vc, "norm_eps", None),
-                default=1e-6,
-            ),
-            # Pixtral / Mistral-3 vision fields
-            model_type=getattr(vc, "model_type", None),
-            head_dim=getattr(vc, "head_dim", None),
-            rope_theta=getattr(vc, "rope_theta", None),
-            # Qwen VL-specific vision fields
-            out_hidden_size=_vis_out_hidden,
-            in_channels=_first_not_none(
-                getattr(vc, "in_channels", None),
-                getattr(vc, "num_channels", None),
-                default=3,
-            ),
-            spatial_merge_size=getattr(vc, "spatial_merge_size", 2),
-            temporal_patch_size=getattr(vc, "temporal_patch_size", 2),
-            num_position_embeddings=getattr(vc, "num_position_embeddings", None),
-            deepstack_visual_indexes=getattr(vc, "deepstack_visual_indexes", None),
-            fullatt_block_indexes=getattr(vc, "fullatt_block_indexes", None),
-            window_size=getattr(vc, "window_size", None),
-            # Gemma4 SigLIP uses clipped linear activations
-            use_clipped_linears=getattr(vc, "use_clipped_linears", False),
-            # Gemma4 SigLIP patch position embedding table size
-            position_embedding_size=getattr(vc, "position_embedding_size", None),
-        )
-    vision_fields["mm_tokens_per_image"] = getattr(vision_source, "mm_tokens_per_image", None)
-    vision_fields["image_token_id"] = getattr(vision_source, "image_token_id", None)
-
-    # MRoPE section — only for composite VL models
-    # (parent_config != config)
-    if parent_config is not None and parent_config is not config:
-        rope_scaling = getattr(config, "rope_scaling", None) or {}
-        rope_parameters = getattr(config, "rope_parameters", None) or {}
-        mrope_section = rope_scaling.get("mrope_section", None) or rope_parameters.get(
-            "mrope_section", None
-        )
-        if mrope_section is not None:
-            vision_fields["mrope_section"] = mrope_section
-
-    # LoRA config (e.g. Phi4-MM)
-    vision_lora = getattr(config, "vision_lora", None)
-    if vision_lora is not None:
-        vision_fields["lora"] = (
-            vision_lora if isinstance(vision_lora, dict) else vars(vision_lora)
-        )
-
-    # Phi4MM image embedding config
-    embd_layer = getattr(config, "embd_layer", None)
-    if isinstance(embd_layer, dict):
-        img_cfg = embd_layer.get("image_embd_layer", {})
-        vision_fields["image_crop_size"] = img_cfg.get("crop_size")
-
-    # TODO(Phase 1): Move phi4mm model-specific logic to a config subclass
-    # override instead of hardcoding here. These SigLIP vision encoder params
-    # are baked into the HF model code and not in the config JSON.
-    if model_type == "phi4mm":
-        vision_fields.update(
-            hidden_size=1152,
-            intermediate_size=4304,
-            num_hidden_layers=27,
-            num_attention_heads=16,
-            image_size=(vision_fields.get("image_crop_size") or 448),
-            patch_size=14,
-            norm_eps=1e-6,
-            image_token_id=getattr(config, "special_image_token_id", 200010),
-        )
-
-    # HunYuan VL-MoT has a flat config.json with no vision_config sub-object.
-    # The vision encoder params are hardcoded (InternViT-style ViT).
-    if model_type == "hunyuan_vl_mot" and not vision_fields.get("hidden_size"):
-        vision_fields.update(
-            hidden_size=1152,
-            intermediate_size=4304,
-            num_hidden_layers=27,
-            num_attention_heads=16,
-            image_size=2048,
-            patch_size=16,
-            norm_eps=1e-6,
-            spatial_merge_size=2,
-            image_token_id=getattr(vision_source, "mask_init_id", 12),
-        )
-
-    # InternVL2 doesn't expose image_token_id in its config — default to
-    # the Qwen2 <IMG_CONTEXT> token id used by InternVL2-* models.
-    parent_model_type = getattr(vision_source, "model_type", None)
-    if parent_model_type in ("internvl_chat", "internvl2", "internvl") or model_type in (
-        "internvl_chat",
-        "internvl2",
-        "internvl",
-    ):
-        if vision_fields.get("image_token_id") is None:
-            vision_fields["image_token_id"] = getattr(
-                vision_source, "img_context_token_id", 151667
-            )
-
-    # Build VisionConfig sub-config if any vision fields are set
-    has_vision = any(
-        v is not None
-        for k, v in vision_fields.items()
-        if k
-        not in (
-            "norm_eps",
-            "in_channels",
-            "spatial_merge_size",
-            "temporal_patch_size",
-        )
-    )
-
-    result: dict = {}
-    if has_vision:
-        result["vision"] = VisionConfig(**vision_fields)
-        # Also set shared fields at top-level for direct access
-        # (e.g. config.image_token_id, config.spatial_merge_size).
-        for shared in (
-            "mm_tokens_per_image",
-            "image_token_id",
-            "spatial_merge_size",
-            "temporal_patch_size",
-            "deepstack_visual_indexes",
-            "fullatt_block_indexes",
-            "window_size",
-            "mrope_section",
-            "image_crop_size",
-        ):
-            val = vision_fields.get(shared)
-            if val is not None:
-                result[shared] = val
-
-    return result
+    return _dispatch(config, parent_config, model_type)
 
 
 def _extract_audio_config(config, parent_config, model_type: str) -> dict:
